+++ src/tests/mocks/SelectWidgetShowCase.tsx
@@ -1,6 +1,0 @@
-import ShowCaseWidgetProps from '../../types/Widgets';
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-export default function SelectWidgetShowCase(_props: ShowCaseWidgetProps) {
-  return <div>SelectWidget</div>;
-}